requirements.txt
blinker==1.5
click==8.1.3
Flask==2.2.2
flask-ldap3-login==0.9.18
flask-requests==0.0.14
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
ldap3==2.9.1
MarkupSafe==2.1.1
pampy==0.3.0
pyasn1==0.4.8
pycryptodome==3.15.0
Werkzeug==2.2.2
WTForms==3.0.1
